Refactor: Flatten folder structure and remove unnecessary subfolders Consolidated the project structure for improved navigation and simplicity. Removed redundant subfolders and organized files directly within their respective main folders. This enhances code readability and streamlines the directory layout.
