+++ JobLeet.WebApi/JobLeet.Core/Entities/Accounts/V1/Admin.cs
@@ -1,5 +1,5 @@
 ﻿using System.ComponentModel.DataAnnotations.Schema;
-namespace JobLeet.WebApi.JobLeet.Core.Entities.Accounts.V1.Users
+namespace JobLeet.WebApi.JobLeet.Core.Entities.Accounts.V1
 {
     public class Admin : BaseEntity
     {
