+++ JobLeet.WebApi/JobLeet.Core/Entities/Accounts/V1/User.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations.Schema;
 
-namespace JobLeet.WebApi.JobLeet.Core.Entities.Accounts.V1.Users
+namespace JobLeet.WebApi.JobLeet.Core.Entities.Accounts.V1
 {
     public class User : BaseEntity
     {
