+++ JobLeet.WebApi/JobLeet.Core/Entities/Common/V1/Education.cs
@@ -1,11 +1,11 @@
-﻿namespace JobLeet.WebApi.JobLeet.Core.Entities.Common.V1.Qualifications
+﻿namespace JobLeet.WebApi.JobLeet.Core.Entities.Common.V1
 {
     public class Education : BaseEntity
     {
         public string Degree { get; set; }
         public string Major { get; set; }
         public string Institution { get; set; }
-        public DateOnly GraduationDate {  get; set; }
-        public Decimal Cgpa {  get; set; }
+        public DateOnly GraduationDate { get; set; }
+        public decimal Cgpa { get; set; }
     }
 }
