+++ JobLeet.WebApi/JobLeet.Core/Entities/Common/V1/PersonName.cs
@@ -1,4 +1,4 @@
-﻿namespace JobLeet.WebApi.JobLeet.Core.Entities.Common.V1.Shared
+﻿namespace JobLeet.WebApi.JobLeet.Core.Entities.Common.V1
 {
     public class PersonName : BaseEntity
     {
