--- JobLeet.WebApi/JobLeet.Core/Entities/Common/V1/Proficiencies/Qualification.cs
@@ -1,16 +1,0 @@
-﻿
-namespace JobLeet.WebApi.JobLeet.Core.Entities.Common.V1.Qualifications
-{
-    public class Qualification : BaseEntity
-    {
-        public List<QualificationType> QualificationTypes {  get; set; }
-        public string QualificationInformation {  get; set; }
-    }
-
-    public enum QualificationType
-    {
-        Education,
-        Skill,
-        Certification
-    }
-}
