+++ JobLeet.WebApi/JobLeet.Core/Entities/Common/V1/Qualification.cs
@@ -1,0 +1,15 @@
+﻿namespace JobLeet.WebApi.JobLeet.Core.Entities.Common.V1
+{
+    public class Qualification : BaseEntity
+    {
+        public List<QualificationType> QualificationTypes { get; set; }
+        public string QualificationInformation { get; set; }
+    }
+
+    public enum QualificationType
+    {
+        Education,
+        Skill,
+        Certification
+    }
+}
