+++ JobLeet.WebApi/JobLeet.Core/Entities/Companies/V1/CompanyProfile.cs
@@ -1,22 +1,21 @@
-﻿using JobLeet.WebApi.JobLeet.Core.Entities.Common.V1.Shared;
-using JobLeet.WebApi.JobLeet.Core.Entities.Companies.V1.IndustryTypes;
+﻿using JobLeet.WebApi.JobLeet.Core.Entities.Common.V1;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace JobLeet.WebApi.JobLeet.Core.Entities.Companies.V1.Profiles
+namespace JobLeet.WebApi.JobLeet.Core.Entities.Companies.V1
 {
     public class CompanyProfile : BaseEntity
     {
         public string ProfileInfo { get; set; }
         [ForeignKey("Address")]
-        public Address CompanyAddress {  get; set; }
+        public Address CompanyAddress { get; set; }
         [ForeignKey("Phone")]
-        public Phone ContactPhone {  get; set; }
+        public Phone ContactPhone { get; set; }
         [ForeignKey("Email")]
-        public Email ContactEmail {  get; set; }
+        public Email ContactEmail { get; set; }
         /// <summary>
         /// Website can be extended in future based on the requirements
         /// </summary>
-        public string Website {  get; set; }
+        public string Website { get; set; }
         [ForeignKey("IndustryType")]
         public IndustryType IndustryType { get; set; }
