+++ JobLeet.WebApi/JobLeet.Core/Entities/Jobs/V1/Application.cs
@@ -1,9 +1,8 @@
-﻿using JobLeet.WebApi.JobLeet.Core.Entities.Common.V1.Shared;
+﻿using JobLeet.WebApi.JobLeet.Core.Entities.Common.V1;
 using JobLeet.WebApi.JobLeet.Core.Entities.Employers.V1;
-using JobLeet.WebApi.JobLeet.Core.Entities.Jobs.V1.Statuses;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace JobLeet.WebApi.JobLeet.Core.Entities.Jobs.V1.Applications
+namespace JobLeet.WebApi.JobLeet.Core.Entities.Jobs.V1
 {
     public class Application : BaseEntity
     {
